+++ src/main/java/com/foxminded/university/service/TimeframeService.java
@@ -1,7 +1,9 @@
 package com.foxminded.university.service;
 
 import java.util.List;
+import java.util.Optional;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -43,7 +45,9 @@ public class TimeframeService {
 
 	@Transactional
 	public void deleteById(Long id) {
-		timeframeDao.deleteById(id);
+		if (isPresentById(id)) {
+			timeframeDao.deleteById(id);
+		}
 	}
 
 	private boolean isTimeframeValid(Timeframe timeframe) {
@@ -51,5 +55,13 @@ public class TimeframeService {
 				&& timeframe.getStartTime() != null
 				&& timeframe.getEndTime() != null
 				&& timeframe.getStartTime().isBefore(timeframe.getEndTime());
+	}
+
+	private boolean isPresentById(Long id) {
+		try {
+			return Optional.of(timeframeDao.findById(id)).isPresent();
+		} catch (EmptyResultDataAccessException exeption) {
+			return false;
+		}
 	}
 }
